Move preprocessors to a separate folder
