+++ pydocmd/preprocessors/simple.py
